requirements.txt
pyinstaller==4.7
firebird-driver==1.4.1